improvements

+++ Code/eyca.invoicing.core/Entities/Project.cs
@@ -11,10 +11,10 @@ namespace eyca.invoicing.core.Entities
     public class Project
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
         [Unique]
         public string Code { get; set; }
-        public int? EmployeeId { get; set; }
+        public long? EmployeeId { get; set; }
         public string EngagementPartnerName { get; set; }
         public string EngagementPartnerEmail { get; set; }
         public string EngagementParnerOfficeAddress { get; set; }
